# app/spec/swagger.yaml
openapi: 3.0.0
info:
  title: flowerstand
  contact:
    name: Tomoyuki Sato
    url: https://mstdn.schoolidol.club/@yazin
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: https://flowerstand.schoolidol.club/api
paths:
...
  /participants:
    get:
      summary: returns all perticipants for a specified flower stand.
      parameters:
      - name: flowerStandId
        in: query
        description: id of flower stand to get participants.
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        "204":
          description: flower stand with specified id does not exists.
        "400":
          description: flowerStandId is not specified or invalid.
    post:
      summary: add a new participant to flower stand.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_4'
        required: true
      responses:
        "200":
          description: OK
        "204":
          description: flower stand with specified id does not exists.
        "401":
          description: invalid participationCode.
        "403":
          description: max number of participants (100) exceeded.
  /participants/{id}:
    get:
      summary: returns a participant.
      parameters:
      - name: id
        in: path
        description: id of participant.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        "204":
          description: participant with specified id does not exists.
    delete:
      summary: delete participant from flower stand.
      parameters:
      - name: id
        in: path
        description: id of participant to delete.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: adminKey
        in: query
        description: administration key issued on create flower stand.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "204":
          description: participant with specified id does not exists.
        "401":
          description: invalid adminKey.
components:
  schemas:
...
    Participant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        twitterId:
          type: string
        instagramId:
          type: string
        facebookId:
          type: string
        mastodonId:
          type: string
...
    body_4:
      required:
      - flowerStandId
      - name
      - participationCode
      type: object
      properties:
        name:
          type: string
          description: participant name.
        flowerStandId:
          type: integer
          description: flower stand id to participate.
        participationCode:
          type: string
          description: participation code of flowerstand.
        twitterId:
          type: string
          description: Twitter id. no leading @.
        instagramId:
          type: string
          description: Instagram id. no leading @.
        facebookId:
          type: string
          description: Facebook id. no leading @.
        mastodonId:
          type: string
          description: Mastodon id. no leading @. ${username}@${host} format.